photo-upload/tests/unit/test_app.py
"""
Unit tests for photo upload Lambda function
"""
import json
import pytest
import base64
from unittest.mock import Mock, patch, MagicMock
from PIL import Image
import io
import sys
import os

# Add the parent directory to sys.path to import the app
sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))))

# Mock the shared imports before importing app
with patch.dict('sys.modules', {
    'shared.utils': MagicMock(),
    'shared.config': MagicMock(),
    'shared.models.photo': MagicMock(),
    'shared.processors.image': MagicMock()
}):
    from app import lambda_handler


class TestPhotoUploadLambdaHandler:
    """Test cases for photo upload Lambda handler"""
    
    def test_lambda_handler_missing_body(self, lambda_context, api_gateway_event):
        """Test handler with missing request body"""
        api_gateway_event['body'] = None
        
        response = lambda_handler(api_gateway_event, lambda_context)
        
        assert response['statusCode'] == 400
        body = json.loads(response['body'])
        assert body['error'] == 'No request body provided'
    
    def test_lambda_handler_invalid_json(self, lambda_context, api_gateway_event):
        """Test handler with invalid JSON body"""
        api_gateway_event['body'] = 'invalid-json'
        
        response = lambda_handler(api_gateway_event, lambda_context)
        
        assert response['statusCode'] == 400
        body = json.loads(response['body'])
        assert 'Invalid JSON' in body['error']
    
    def test_lambda_handler_missing_required_fields(self, lambda_context, api_gateway_event):
        """Test handler with missing required fields"""
        api_gateway_event['body'] = json.dumps({
            'entity_type': 'user'
            # Missing entity_id, photo_type, image
        })
        
        response = lambda_handler(api_gateway_event, lambda_context)
        
        assert response['statusCode'] == 400
        body = json.loads(response['body'])
        assert 'Missing required field' in body['error']
    
    def test_lambda_handler_invalid_entity_type(self, lambda_context, valid_photo_upload_event):
        """Test handler with invalid entity type"""
        body_data = json.loads(valid_photo_upload_event['body'])
        body_data['entity_type'] = 'invalid'
        valid_photo_upload_event['body'] = json.dumps(body_data)
        
        response = lambda_handler(valid_photo_upload_event, lambda_context)
        
        assert response['statusCode'] == 400
        body = json.loads(response['body'])
        assert 'Invalid entity_type' in body['error']
    
    def test_lambda_handler_invalid_photo_type(self, lambda_context, valid_photo_upload_event):
        """Test handler with invalid photo type"""
        body_data = json.loads(valid_photo_upload_event['body'])
        body_data['photo_type'] = 'invalid'
        valid_photo_upload_event['body'] = json.dumps(body_data)
        
        response = lambda_handler(valid_photo_upload_event, lambda_context)
        
        assert response['statusCode'] == 400
        body = json.loads(response['body'])
        assert 'Invalid photo_type' in body['error']
    
    def test_lambda_handler_invalid_image_format(self, lambda_context, valid_photo_upload_event):
        """Test handler with invalid image format"""
        body_data = json.loads(valid_photo_upload_event['body'])
        body_data['image'] = 'data:text/plain;base64,aGVsbG8='  # text/plain instead of image
        valid_photo_upload_event['body'] = json.dumps(body_data)
        
        response = lambda_handler(valid_photo_upload_event, lambda_context)
        
        assert response['statusCode'] == 400
        body = json.loads(response['body'])
        assert 'Invalid image format' in body['error']
    
    @patch('app.cleanup_existing_photos')
    @patch('app.process_and_upload_image')
    @patch('app.Photo')
    def test_lambda_handler_success(
        self, 
        mock_photo_class, 
        mock_process_upload, 
        mock_cleanup,
        lambda_context, 
        valid_photo_upload_event,
        sample_photo_record
    ):
        """Test successful photo upload"""
        # Setup mocks
        mock_cleanup.return_value = {'deleted_count': 1}
        mock_process_upload.return_value = {
            'photo_id': 'test-photo-123',
            'urls': {
                'thumbnail': 'https://bucket/thumbnail.jpg',
                'standard': 'https://bucket/standard.jpg',
                'high_res': 'https://bucket/high_res.jpg'
            },
            'metadata': {'file_size': 1024}
        }
        
        mock_photo_instance = Mock()
        mock_photo_instance.to_dict.return_value = sample_photo_record
        mock_photo_class.return_value = mock_photo_instance
        
        response = lambda_handler(valid_photo_upload_event, lambda_context)
        
        assert response['statusCode'] == 200
        body = json.loads(response['body'])
        assert body['success'] is True
        assert 'photo_id' in body
        assert 'urls' in body
    
    def test_lambda_handler_exception_handling(self, lambda_context, valid_photo_upload_event):
        """Test exception handling in Lambda handler"""
        with patch('app.process_and_upload_image', side_effect=Exception('Test error')):
            response = lambda_handler(valid_photo_upload_event, lambda_context)
            
            assert response['statusCode'] == 500
            body = json.loads(response['body'])
            assert body['error'] == 'Internal server error'


class TestImageProcessing:
    """Test cases for image processing functions"""
    
    def create_test_image(self, format='JPEG', size=(100, 100), color='RGB'):
        """Create a test image for testing"""
        img = Image.new(color, size, color=(255, 255, 255))
        img_bytes = io.BytesIO()
        img.save(img_bytes, format=format)
        img_bytes.seek(0)
        return img_bytes.getvalue()
    
    def test_parse_data_uri_valid_jpeg(self):
        """Test parsing valid JPEG data URI"""
        # This would need to be implemented when we refactor the actual functions
        pass
    
    def test_parse_data_uri_invalid_format(self):
        """Test parsing invalid data URI format"""
        pass
    
    def test_image_resizing_and_processing(self):
        """Test image resizing and processing"""
        pass


class TestS3Operations:
    """Test cases for S3 operations"""
    
    @patch('app.s3_client')
    def test_s3_upload_success(self, mock_s3):
        """Test successful S3 upload"""
        pass
    
    @patch('app.s3_client')
    def test_s3_upload_failure(self, mock_s3):
        """Test S3 upload failure handling"""
        pass


class TestDynamoDBOperations:
    """Test cases for DynamoDB operations"""
    
    def test_photo_record_creation(self, mock_dynamodb):
        """Test photo record creation in DynamoDB"""
        pass
    
    def test_photo_record_update(self, mock_dynamodb):
        """Test photo record update in DynamoDB"""
        pass


if __name__ == '__main__':
    pytest.main([__file__])