requirements-test.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
moto[dynamodb,s3]==4.2.14
boto3-stubs[essential]==1.34.0
mypy==1.7.1
black==23.11.0
flake8==6.1.0